refactor: data layer분리
