+++ lib/features/route_planning/application/services/polyline_utils.dart
@@ -1,6 +1,6 @@
 import 'package:google_polyline_algorithm/google_polyline_algorithm.dart';
 import 'package:latlong2/latlong.dart';
-import 'package:ridingmate/features/route_planning/data/models/route_data.dart';
+import 'package:ridingmate/features/route_planning/domain/entities/route_data.dart';
 
 class PolylineUtils {
   static String encodeRouteSegments(List<RouteData> routeSegments) {
