+++ lib/features/route_planning/data/models/route_api_response_model.dart
@@ -1,0 +1,35 @@
+class RouteApiResponseModel {
+  const RouteApiResponseModel({
+    required this.coordinates,
+    required this.distance,
+    required this.duration,
+    required this.rawAscent,
+    required this.rawDescent,
+  });
+
+  factory RouteApiResponseModel.fromJson(Map<String, dynamic> json) {
+    final List<List<dynamic>> coordinates =
+        (json['features'][0]['geometry']['coordinates'] as List<dynamic>)
+            .cast<List<dynamic>>();
+
+    final Map<String, dynamic> properties =
+        json['features'][0]['properties'] as Map<String, dynamic>;
+    final Map<String, dynamic> summary =
+        properties['summary'] as Map<String, dynamic>;
+
+    return RouteApiResponseModel(
+      coordinates: coordinates,
+      distance: (summary['distance'] as num).toDouble(),
+      duration: (summary['duration'] as num).toDouble(),
+      rawAscent: (properties['ascent'] as num?)?.toDouble() ?? 0.0,
+      rawDescent: (properties['descent'] as num?)?.toDouble() ?? 0.0,
+    );
+  }
+
+  /// 원시 좌표 데이터 [longitude, latitude, elevation]
+  final List<List<dynamic>> coordinates;
+  final double distance; // 단위 : 미터
+  final double duration; // 단위 : 초
+  final double rawAscent; // ORS 원시 총 상승고도(미터)
+  final double rawDescent; // 총 하강고도(미터)
+}
